+++ packages/compiler-cli/src/ngtsc/diagnostics/index.ts
@@ -6,6 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
+export {COMPILER_ERRORS_WITH_GUIDES, ERROR_DETAILS_PAGE_BASE_URL} from './src/docs';
 export {FatalDiagnosticError, isFatalDiagnosticError, makeDiagnostic, makeRelatedInformation} from './src/error';
-export {COMPILER_ERRORS_WITH_GUIDES, ERROR_DETAILS_PAGE_BASE_URL, ErrorCode, ngErrorCode} from './src/error_code';
-export {replaceTsWithNgInErrors} from './src/util';
+export {ErrorCode} from './src/error_code';
+export {ngErrorCode, replaceTsWithNgInErrors} from './src/util';
